+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This change log follows the conventions of
 
 ## [Unreleased][unreleased]
 
+Nothing so far.
+
+## [0.1.4] - 2016-05-18
+
 ### Fixed
 
 - Accept shorter, 208-byte CDJ status packets sent by non-nexus players.
@@ -88,7 +92,8 @@ This change log follows the conventions of
 - Intial early release of DeviceFinder.
 
 
-[unreleased]: https://github.com/brunchboy/beat-link/compare/v0.1.3...HEAD
+[unreleased]: https://github.com/brunchboy/beat-link/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/brunchboy/beat-link/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/brunchboy/beat-link/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/brunchboy/beat-link/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/brunchboy/beat-link/compare/v0.1.0...v0.1.1